+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#242](https://github.com/os2display/display-api-service/pull/242)
+  - Changed calendar api feed type to use display name from resources feed instead of events feed.
+
 ## [2.5.0] - 2025-05-09
 
 - [#240](https://github.com/os2display/display-api-service/pull/240)